requirements.txt
requests
python-dotenv
aiohttp
pydantic
numpy
scikit-learn
pandas
torch
sentence-transformers
rich
alive-progress
colorama
